+++ ASICamera.py
@@ -4,19 +4,25 @@ import time
 from CameraInterface import CameraInterface
 import zwoasi as asi
 import Display
+from typing import Dict
+import logging
+import utils
 
 
 class ASICamera(CameraInterface):
     """The camera class for ASI cameras.  Implements the CameraInterface interface."""
 
-    def __init__(self, handpad: Display) -> None:
+    def __init__(self, handpad: Display, images_path=Path('/dev/shm/images'), home_path=Path.cwd()) -> None:
         """Initializes the ASI camera
 
         Parameters:
         handpad (Display): The link to the handpad"""
 
-        self.home_path = str(Path.home())
         self.handpad = handpad
+        self.images_path: Path = images_path
+        self.home_path: Path = home_path 
+        self.stills_path: Path = home_path / "Stills"
+        utils.create_path(self.stills_path) # create stills dir if not already therew
 
         # find a camera
         asi.init("/lib/zwoasi/armv7/libASICamera2.so")
@@ -24,7 +30,7 @@ class ASICamera(CameraInterface):
         if num_cameras == 0:
             self.handpad.display("Error:", " no camera found", "")
             self.camType = "not found"
-            print("camera not found")
+            logging.info("camera not found")
             time.sleep(1)
             exit()
         else:
@@ -33,7 +39,7 @@ class ASICamera(CameraInterface):
             camera_id = 0
             self.initialize()
             self.handpad.display("ZWO camera found", "", "")
-            print("ZWO camera found")
+            logging.info("ZWO camera found")
             time.sleep(1)
 
     def initialize(self) -> None:
@@ -54,7 +60,7 @@ class ASICamera(CameraInterface):
         camera.set_image_type(asi.ASI_IMG_RAW8)
 
     def capture(
-        self, exposure_time: float, gain: float, radec: str, m13: bool, polaris: bool
+            self, exposure_time: float, gain: float, radec: str, extras: Dict
     ) -> None:
         """Capture an image with the camera
 
@@ -72,23 +78,10 @@ class ASICamera(CameraInterface):
         timestr = time.strftime("%Y%m%d-%H%M%S")
         camera.set_control_value(asi.ASI_GAIN, gain)
         camera.set_control_value(asi.ASI_EXPOSURE, exposure_time)  # microseconds
-
-        if m13 == True:
-            copyfile(
-                self.home_path + "/Solver/test.jpg",
-                self.home_path + "/Solver/images/capture.jpg",
-            )
-        elif polaris == True:
-            copyfile(
-                self.home_path + "/Solver/polaris.jpg",
-                self.home_path + "/Solver/images/capture.jpg",
-            )
-            print("using Polaris")
-        else:
-            camera.capture(filename=self.home_path + "/Solver/images/capture.jpg")
-            copyfile(
-                self.home_path + "/Solver/images/capture.jpg",
-                self.home_path + "/Solver/Stills/" + timestr + "_" + radec + ".jpg",
+        capture_path = self.images_path / "capture.jpg"
+        camera.capture(filename=capture_path)
+        copyfile( capture_path,
+                self.stills_path / f"{timestr}_{radec}.jpg",
             )
 
         return