+++ QHYCamera.py
@@ -2,23 +2,26 @@ from pathlib import Path
 from shutil import copyfile
 import time
 from CameraInterface import CameraInterface
-import zwoasi as asi
 import Display
 import cv2
 import qhyccd
 from ctypes import *
-
+from typing import Dict
+import utils
 
 class QHYCamera(CameraInterface):
     """The camera class for ZWO cameras.  Implements the CameraInterface interface."""
 
-    def __init__(self, handpad: Display) -> None:
+    def __init__(self, handpad: Display, images_path=Path('/dev/shm/images'), home_path=Path.cwd()) -> None:
         """Initializes the QHY camera
 
         Parameters:
         handpad (Display): The link to the handpad"""
 
-        self.home_path = str(Path.home())
+        self.home_path = home_path 
+        self.images_path = images_path 
+        self.stills_path: Path = home_path / "Stills"
+        utils.create_path(self.stills_path) # create stills dir if not already therew
         self.handpad = handpad
         self.camType = "QHY"
         self.initialize()
@@ -35,7 +38,7 @@ class QHYCamera(CameraInterface):
         print('Found camera:',ident)
 
     def capture(
-        self, exposure_time: float, gain: float, radec: str, m13: bool, polaris: bool
+            self, exposure_time: float, gain: float, radec: str, extra: Dict 
     ) -> None:
         """Capture an image with the camera
 
@@ -54,25 +57,13 @@ class QHYCamera(CameraInterface):
         camera.SetGain(gain)
         camera.SetExposure(exposure_time/1000)  # milliseconds
 
-        if m13 == True:
-            copyfile(
-                self.home_path + "/Solver/test.jpg",
-                self.home_path + "/Solver/images/capture.jpg",
-            )
-        elif polaris == True:
-            copyfile(
-                self.home_path + "/Solver/polaris.jpg",
-                self.home_path + "/Solver/images/capture.jpg",
-            )
-            print("using Polaris")
-        else:
-            img = camera.GetSingleFrame()
-            cv2.imwrite(self.home_path + "/Solver/images/capture.jpg",img)
-            copyfile(
-                self.home_path + "/Solver/images/capture.jpg",
-                self.home_path + "/Solver/Stills/" + timestr + "_" + radec + ".jpg",
-            )
-
+        img = camera.GetSingleFrame()
+        capture_path = self.images_path / "capture.jpg"
+        cv2.imwrite(capture_path,img)
+        copyfile(
+            capture_path,
+            self.stills_path / f"{timestr}_{radec}.jpg",
+        )
         return
 
     def get_cam_type(self) -> str: